06-mesop-search/requirements.txt
mesop==0.8.0
google-api-core==2.19.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.11.13
